refactor(#172): restructure applied/theming/ → applied/hmi/{theming,surfaces,visualization,input,report,explorer}/
The applied/theming/ directory had accumulated 13 files across concerns
well beyond color schemes: render surfaces, visualization grammar,
input/keybindings, reports, self-referential explorer. Only 5 of the
files were actually about theming (base16, palette, ontology, schemes,
variants, theme_package).

This restructure splits them into their honest homes under a new
applied/hmi/ (Human-Machine Interface) umbrella:

  applied/hmi/
  ├── theming/       base16 + ontology + schemes + variants + theme_package
  ├── surfaces/      (was theming/surfaces.rs) render targets
  ├── visualization/ (was theming/visualization.rs) Bertin/Cleveland/Munzner/Wickham
  ├── input/         (was theming/keybindings.rs + modes.rs)
  ├── report/        (was theming/report.rs + report_spec.rs + validate_themes.rs)
  │                  → renamed spec.rs, generator.rs, validator.rs
  └── explorer/      (was theming/explorer.rs + shader_params.rs)

"HMI" is the right parent — defense, aerospace, medical devices,
industrial PLCs all use "HMI" as the standard term for this layer.
"UX" is product-design language; "HMI" is interface-engineering
language, which matches pr4xis's target market.

Cross-sub-directory imports rewired:
- hmi/surfaces/ontology.rs       → crate::applied::hmi::theming::{base16, ontology}
- hmi/report/spec.rs             → crate::applied::hmi::visualization::ontology
- hmi/report/validator.rs        → crate::applied::hmi::theming::{base16, ontology}
- hmi/explorer/ontology.rs       → crate::applied::hmi::report::generator
- hmi/theming/theme_package.rs   → crate::applied::hmi::theming::base16 (test)

Intra-cluster imports (all super::*) stay unchanged because each
cluster moved together.

No external consumers of applied::theming::* existed outside the
directory itself, so this restructure has zero cross-crate impact.

README.md, citings.md, and papers/ (with the 3 cited PDFs) move with
the theming cluster to applied/hmi/theming/. Follow-up PR will add
README.md + citings.md for the new sibling sub-ontologies (surfaces,
visualization, input, report, explorer).

Verified:
- cargo build -p pr4xis-domains (clean)
- cargo test --workspace --lib (all green)
- cargo test -p pr4xis-domains --lib (4371 passed, 17 ignored)
- cargo test -p pr4xis-domains test_scan_all_datasets -- --nocapture
  regenerates the identical 464-theme report from its new location
  (applied::hmi::report::validator::tests::test_scan_all_datasets)
