+++ crates/domains/src/applied/hmi/explorer/mod.rs
@@ -1,0 +1,7 @@
+/// Explorer — self-referential visualization of reasoning traces.
+///
+/// The explorer visualizes the ontology using the ontology's own theme.
+/// Concept nodes light up as axioms evaluate, colored by the active theme.
+/// Shader params expose GPU-side rendering controls.
+pub mod ontology;
+pub mod shader_params;
