+++ crates/domains/src/applied/hmi/explorer/shader_params.rs
