+++ crates/domains/src/applied/hmi/input/keybindings.rs
