+++ crates/domains/src/applied/hmi/input/modes.rs
