+++ crates/domains/src/applied/hmi/theming/base16.rs
