+++ crates/domains/src/applied/hmi/theming/ontology.rs
