+++ crates/domains/src/applied/hmi/theming/papers/draft-color-theming-ontology.md
