+++ crates/domains/src/applied/hmi/theming/papers/harel-statecharts-1987.pdf
