+++ crates/domains/src/applied/hmi/theming/schemes.rs
