+++ crates/domains/src/applied/hmi/theming/variants.rs
