+++ crates/domains/src/applied/hmi/visualization/mod.rs
@@ -1,0 +1,6 @@
+/// Visualization — formal model of visual encoding and perception.
+///
+/// Bertin's 7 visual variables, Cleveland-McGill perceptual accuracy ranking,
+/// Munzner channel effectiveness, Shneiderman's overview-zoom-filter-detail
+/// mantra, Wickham's layered grammar of graphics.
+pub mod ontology;
