+++ crates/domains/src/applied/hmi/visualization/ontology.rs
