+++ crates/domains/src/applied/mod.rs
@@ -1,12 +1,12 @@
 /// Applied sciences — engineering and technology.
 /// DOLCE: Process.
 pub mod hardware;
+pub mod hmi;
 pub mod industrial;
 pub mod localization;
 pub mod navigation;
 pub mod perception;
 pub mod sensor_fusion;
 pub mod space;
-pub mod theming;
 pub mod tracking;
 pub mod underwater;
